+++ src/app/shared/components/web-footer.ts
@@ -17,7 +17,7 @@ export class WebFooterComponent implements OnInit {
   constructor() {
     this.model = {
       sdkVersion: OJP.SDK_VERSION,
-      lastUpdate: '17.September 2024',
+      lastUpdate: '03.October 2024',
     }
   }