+++ src/app/station-board/search/station-board-search.component.ts
@@ -413,7 +413,9 @@ export class StationBoardSearchComponent implements OnInit {
 
   public showRequestXmlPopover() {
     const dialogRef = this.debugXmlPopover.open(DebugXmlPopoverComponent, {
-      position: { top: '10px' },
+      position: { top: '20px' },
+      width: '50vw',
+      height: '90vh',
     });
     dialogRef.afterOpened().subscribe(() => {
       if (this.currentRequestInfo) {